fix floats on db

+++ server/db/db.js
@@ -43,7 +43,7 @@ var Income = sequelize.define('income', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
     account: {
@@ -83,7 +83,7 @@ var Expense = sequelize.define('expense', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
     comment: {
@@ -115,7 +115,7 @@ var Savings = sequelize.define('savings', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
     comment: {
@@ -147,7 +147,7 @@ var Invest = sequelize.define('invest', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
     comment: {
@@ -211,7 +211,7 @@ var CurrentTotalIncome = sequelize.define('currenttotalincome', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
   }, {
@@ -227,7 +227,7 @@ var CurrentTotalExpenses = sequelize.define('currenttotalexpenses', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
   }, {
@@ -243,7 +243,7 @@ var CurrentTotalSavings = sequelize.define('currenttotalsavings', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
   }, {
@@ -259,7 +259,7 @@ var CurrentTotalInvest = sequelize.define('currenttotalinvest', {
       autoIncrement: true
     },
     amount: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.FLOAT(10,2),
       allowNull: false
     },
   }, {
